+++ Localization.h
@@ -1,0 +1,14 @@
+﻿
+static struct {
+	char const* language;
+	wchar_t const* done;
+	wchar_t const* erase;
+	wchar_t const* highlight;
+	wchar_t const* pen;
+	wchar_t const* title;
+} localization[] {
+	{ "en-US", L"Done", L"Erase", L"Highlight", L"Pen", L"Snipping Tool", },
+	{ "fr-FR", L"Terminé", L"Effacer", L"Surligner", L"Stylo", L"Outil Capture", },
+	{ "ja-JP", L"完了", L"消去する", L"ハイライト", L"ペン", L"タイトル", },
+};
+
