+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "preinstall": "npm run build",
-    "build": "cl ScreenSnippet.cpp /O2 /MT /nologo"
+    "build": "cl ScreenSnippet.cpp /O2 /MT /nologo /link /SUBSYSTEM:WINDOWS"
   },
   "repository": {
     "type": "git",
